Resolve ONLY_FULL_GROUP_BY issue in map view
MySQL, beginning with version 5.7, comes with a new default for
SQL_MODE. The new default requires to be more explicit when selecting
non-functionally-dependent columns in a grouped query.

The previous grouping was used to limit markers per spot on the map
(more than one marker per location isn't visible anyway). However, this
results in functionally-independent selects, since there _are_ some
places in the database that map to the different spot, but have
different names.

Dagmar will fix those places in the input data - I will fall back to
grouping by place and ignoring (not displaying) multiple markers per
coordinate.

If neccessary even after data repair, this issue could be resolved by
checking in PHP and slightly randomising marker positions in this case.
Or on the client side in the marker library?