Fix previous group by fix
The fix to simply add "r." to the group clause was faulty, because "id"
did not mean the database value, but the calculated column "id" from the
IF clause. This lead to duplicated results.

To counteract this, I reversed the conditions and select distinct
person IDs from a UNION statement in the WHERE clause.